Add `recode` to README.md
Add `recode` to README.md.

The `recode` program is a handy front-end to the recode library that converts files between character sets and usages. It recognizes or produces over 200 different character sets (or about 300 if combined with an `iconv` library) and transliterates files between almost any pair. When exact transliteration are not possible, it gets rid of offending characters or falls back on approximations.

More information at: https://github.com/rrthomas/recode/